requirements.txt
absl-py==1.4.0
altgraph==0.17.4
apiclient==1.0.4
array-record==0.4.1
astunparse==1.6.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dm-tree==0.1.8
etils==1.8.0
flatbuffers==24.3.25
fsspec==2024.3.1
gast==0.4.0
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.62.1
gspread==6.1.0
h5py==3.11.0
httplib2==0.22.0
idna==3.6
importlib_resources==6.4.0
joblib==1.4.0
keras==2.13.1
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
numpy==1.24.3
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.10.0.82
opt-einsum==3.3.0
packaging==24.0
pandas==2.0.3
pefile==2023.2.7
Pillow==10.0.0
promise==2.3
protobuf==3.20.3
psutil==5.9.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyDrive==1.3.1
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.10.1
six==1.16.0
StrEnum==0.4.15
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.14.0
termcolor==2.4.0
threadpoolctl==3.4.0
toml==0.10.2
tqdm==4.66.2
typing_extensions==4.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.0
Werkzeug==3.0.2
wrapt==1.16.0
zipp==3.18.1
